add: content with id for navigation
